Split parts in smaller components
