+++ client/blocks/importer/wordpress/upgrade-plan/types.ts
@@ -9,3 +9,7 @@ export type HostingDetailsItem = {
 export type HostingDetails = {
 	[ key: string ]: HostingDetailsItem;
 };
+
+export type UpgradePlanDetailsProps = {
+	children: React.ReactNode;
+};
